Consume now returns stop func
