+++ kafka/consumer/package_test.go
@@ -9,22 +9,30 @@ import (
 
 func TestPackage(t *testing.T) {
 	ctx := context.Background()
-	ch := Consume(ctx, "topic-name")
+	ch, stop := Consume(ctx, "topic-name")
+	// when finished close it
+	defer func() {
+		err := stop()
+		if err != nil {
+			fmt.Printf("Error: %v", err)
+		}
+	}()
 
 	// Read the next message from the topic
 	msg, ok := <-ch
 	fmt.Printf("Channel open=%t, message=%v\n", ok, msg)
-
-	// when finished close it
-	err := Stop("topic-name")
-	if err != nil {
-		fmt.Printf("Error: %v", err)
-	}
 }
 
 func TestPackageWithTimeout(t *testing.T) {
 	ctx := context.Background()
-	ch := Consume(ctx, "topic-name")
+	ch, stop := Consume(ctx, "topic-name")
+	// when finished close it
+	defer func() {
+		err := stop()
+		if err != nil {
+			fmt.Printf("Error: %v", err)
+		}
+	}()
 
 	select {
 	case msg, ok := <-ch:
@@ -33,12 +41,6 @@ func TestPackageWithTimeout(t *testing.T) {
 	case <-time.After(time.Duration(1) * time.Second):
 		fmt.Println("No message received before timeout")
 	}
-
-	// when finished close it
-	err := Stop("topic-name")
-	if err != nil {
-		fmt.Printf("Error: %v", err)
-	}
 }
 
 func TestPackageWithDeadline(t *testing.T) {
@@ -46,7 +48,14 @@ func TestPackageWithDeadline(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), deadline)
 	defer cancel()
 
-	ch := Consume(ctx, "topic-name")
+	ch, stop := Consume(ctx, "employee.v1")
+	// when finished close it
+	defer func() {
+		err := stop()
+		if err != nil {
+			fmt.Printf("Error: %v", err)
+		}
+	}()
 
 	ok := true
 	for ok {
@@ -58,11 +67,5 @@ func TestPackageWithDeadline(t *testing.T) {
 			fmt.Println("Context deadline received. Stopping.")
 			ok = false
 		}
-	}
-
-	// when finished close it
-	err := Stop("topic-name")
-	if err != nil {
-		fmt.Printf("Error: %v", err)
 	}
 }
